requirements.txt
requests>=2.31.0
websockets>=12.0
websocket-client>=1.6.0
pydantic>=2.8.2
PyYAML>=6.0.2
PyNaCl>=1.5.0
